% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_variable.R
\name{cluster_variable}
\alias{cluster_variable}
\title{Cluster variable 1 based on mean values. Pass columns directly (not as strings)
e.g. df \%>\% cluster_variable(drugname, gene, expression) \%>\% ggplot() Clusters by first variable only, so run twice to cluster both ways}
\usage{
cluster_variable(
  df,
  cluster_var1,
  cluster_var2,
  value_var,
  method = "euclidean"
)
}
\arguments{
\item{df}{data.frame of values}

\item{cluster_var1}{Column to cluster}

\item{cluster_var2}{Column not to be clustered}

\item{value_var}{Column with values to be clustered by}

\item{method}{Clustering method}
}
\value{
data.frame with clustered column as reordered factor
}
\description{
Cluster variable 1 based on mean values. Pass columns directly (not as strings)
e.g. df \%>\% cluster_variable(drugname, gene, expression) \%>\% ggplot() Clusters by first variable only, so run twice to cluster both ways
}
