% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactive_data_select.R
\name{interactive_data_select}
\alias{interactive_data_select}
\title{Interactively select intensity data using a lasso gate}
\usage{
interactive_data_select(data, channel_1, channel_2)
}
\arguments{
\item{data}{data}

\item{channel_1}{channel 1}

\item{channel_2}{channel 2}
}
\value{
Coordinates for gate
}
\description{
Interactively select intensity data using a lasso gate
}
