% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_confusion.R
\name{plot_confusion}
\alias{plot_confusion}
\title{Plot a nicely formatted model confusion matrix}
\usage{
plot_confusion(confusion_table)
}
\arguments{
\item{confusion_table}{The confusion table from the trained model}
}
\description{
Plot a nicely formatted model confusion matrix
}
