% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_intensity_2d.R
\name{plot_intensity_2d}
\alias{plot_intensity_2d}
\title{Interactive plot of 2d intensity data}
\usage{
plot_intensity_2d(
  data,
  channel_1,
  channel_2,
  bins = 500,
  plot_all = FALSE,
  plotly = T
)
}
\arguments{
\item{data}{data}

\item{channel_1}{channel 1}

\item{channel_2}{channel 2}

\item{bins}{Number of bins}

\item{plot_all}{If FALSE, takes subsample of 500,000 data points to plot}

\item{plotly}{Make plot interactive using plotly}
}
\description{
Interactive plot of 2d intensity data
}
